I just uploaded my jupyter notebook for the project
The question that I was trying to answer is which economic factors are most strongly correlated with the quality of life ranking of a country? I went through the data science pipeline in order to arrive at an answer. I used requests and BeautifulSoup to scrape the web and collect the data, pandas for data cleaning, seaborn and matplotlib for generating data visualizations and scikit-learn to model the data. All of my work can be found in this jupyter notebook.